requirements.txt
pandas==1.3.5
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
sqlalchemy==1.4.25
python-dotenv==0.19.2
mysql-connector-python==8.1.0
